VerdantGuard_Model_Api_Flask/app/requirements.txt
flask==3.0.3
tensorflow==2.17.0
Pillow==11.0.0
numpy==1.26.4
streamlit==3.1
